simplify UIPath specific JSON feature
- expose the obvious reason why it was created and re-use as much as
possible from existing odata stack.
